requirements.txt
python-dotenv==0.21.0
fastapi==0.86.0
uvicorn==0.18.3
requests-oauthlib==1.3.1